+++ PHP-Build/src/Plugswork/Module/LogModule.php
@@ -16,7 +16,8 @@ use Plugswork\Plugswork;
 class LogModule{
     
     private $plugin;
-    private $enableLog, $logsFolder, $unifyLog;
+    private $enableLog = [];
+    private $logsFolder, $unifyLog;
     
     public function __construct(Plugswork $plugin){
         $this->plugin = $plugin;
@@ -34,11 +35,7 @@ class LogModule{
         }else{
             $this->unifyLog = false;
         }
-        if(isset($st["enableLog"])){
-            $this->enableLog = true;
-        }else{
-            $this->enableLog = false;
-        }
+        $this->enableLog = $st["enable"];
         $this->settings = $st;
         $this->loadCache();
     }
@@ -55,13 +52,13 @@ class LogModule{
         }
     }
     
-    public function write($log, $warning = false, $type = null){
-        if($this->enableLog){
+    public function write($log, $warning = false, $type = "main"){
+        if(isset($this->enableLog[$type])){
             $warn = "";
             if($warning){
                 $warn = "[!] ";
             }
-            if($this->unifyLog || $type == null){
+            if($this->unifyLog || $type == "main"){
                 file_put_contents($this->logsFolder."main.log", $warn.date("[Y-m-d H:i:s] ").$log."\n", FILE_APPEND);
             }else{
                 file_put_contents($this->logsFolder.$type.".log", $warn.date("[Y-m-d H:i:s] ").$log."\n", FILE_APPEND);