+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+####6.php.bleed (4.php.bleed Patch II)
+[NEW]
+- getInstance for developer
+[IMPROVEMENTS]
+- Improved /bc broadcast command
+[FIXES]
+- Fix typo caused settings unable to load
+
+
 ####5.php.bleed (4.php.bleed Patch)
 [FIXES]
 - Fix 'Unkown Command' while registering console language